+++ R/ReadNetCDF.R
@@ -1,6 +1,6 @@
 #' Read NetCDF files.
 #'
-#' Using the \code{\link[ncdf4]{ncdf4-package}} package, it reads an .nc file. The advantage
+#' Using the \code{\link[ncdf4]{ncdf4-package}} package, it reads a NetCDF file. The advantage
 #' over using \code{\link[ncdf4]{ncvar_get}} is that the output is a tidy data.table
 #' with proper dimensions.
 #'
@@ -8,15 +8,24 @@
 #' @param vars a character vector with the name of the variables to read. If
 #' \code{NULL}, then it read all the variables.
 #' @param out character indicating the type of output desired
+#' @param subset a named list a subseting vectors. See Details.
+#'
+#' @details
+#' `subset` must be a named list for subsetting. Names must match dimensions
+#' specified in the NetCDF file and each element must be a vector of the same tipe
+#' of that dimension whose range will be use for subsetting. You don't need to
+#' provide and exact range that matches the actual gridpoints of the file; the
+#' closest gridpoint will be selected.
 #'
 #' @return
 #' The return format is specified by `out`. It can be a data table in which each
 #' column is a variable and each row, an observation; an array with named
-#' dimensions; or a vector. Either of these two options are much faster than the
+#' dimensions; or a vector. Since it's possible to return multiple arrays or
+#' vectors (one for each variable), for consistency the return type is always a
+#' list. Either of these two options are much faster than the
 #' first since the most time consuming part is the melting of the array
 #' returned by [ncdf4::ncvar_get]. `out = "vector"` is particularly useful for
-#' adding new variables to an existing data frame with the same dimensions. Note
-#' that only one variable can be retrieved at a time with these formats.
+#' adding new variables to an existing data frame with the same dimensions.
 #'
 #' Finally, it can also be `vars`, in which case it returns a list with the name
 #' of the available variables and the dimensions of the spatiotemporal grid.
@@ -35,13 +44,14 @@
 #' field <- ReadNetCDF(file, vars = c(var = variables$vars[1]))
 #' # Add a new variable.
 #' # ¡Make sure it's on the same exact grid!
-#' field[, var2 := ReadNerCDF(file2, out = "vector")]
+#' field[, var2 := ReadNerCDF(file2, out = "vector", subset = list(lat = 90:10))]
 #' }
 #'
 #' @export
 #' @importFrom lubridate years weeks days hours minutes seconds milliseconds ymd_hms
 #' @import data.table udunits2
-ReadNetCDF <- function(file, vars = NULL, out = c("data.frame", "vector", "array", "vars")) {
+ReadNetCDF <- function(file, vars = NULL, out = c("data.frame", "vector", "array", "vars"),
+                       subset = NULL) {
     ncfile <- ncdf4::nc_open(file)
 
     if (is.null(vars)) {
@@ -89,13 +99,50 @@ ReadNetCDF <- function(file, vars = NULL, out = c("data.frame", "vector", "array
         return(r)
     }
 
+    ## Hago los subsets
+    # Me fijo si faltan dimensiones
+    subset.extra <- names(subset)[!(names(subset) %in% names(dimensions))]
+    if (length(subset.extra) != 0) {
+        stop(paste0("Subseting dimensions not found: ",
+                       paste0(subset.extra, collapse = ", "), "."))
+    }
+
     # Leo las variables y las meto en una lista.
     nc <- list()
+
     dim.length <- vector("numeric", length = length(vars))
     for (v in seq_along(vars)) {
-        var1 <- ncdf4::ncvar_get(ncfile, vars[v], collapse_degen = FALSE)
+
+        # Para cada variable, veo start y count
         order <- ncfile$var[[vars[v]]]$dimids
-        dimnames(var1) <- dimensions[dims[as.character(order)]]
+        start <- rep(1, length(order))
+        names(start) <- names(dimensions[dims[as.character(order)]])
+        count <- rep(-1, length(order))
+        names(count) <- names(dimensions[dims[as.character(order)]])
+
+        sub.dimensions <- dimensions
+
+        for (s in names(subset)[names(subset) %in% names(start)]) {
+            d <- dimensions[[s]]
+            sub <- subset[[s]]
+
+            if (.is.somedate(sub[1])) {
+                start[[s]] <- which(lubridate::as_datetime(d) %~% min(sub))
+                count[[s]] <- abs(which(lubridate::as_datetime(d) %~% max(sub) ) - start[[s]] )
+            } else {
+                start[[s]] <- which(d %~% min(sub))
+                count[[s]] <- abs(which(d %~% max(sub)) - start[[s]])
+            }
+
+            if(count[[s]] == 0) count[[s]] <- 1
+
+            sub.dimensions[[s]] <- dimensions[[s]][seq.int(start[[s]], start[[s]] + count[[s]] - 1)]
+        }
+
+        var1 <- ncdf4::ncvar_get(ncfile, vars[v], collapse_degen = FALSE, start = start,
+                                 count = count)
+
+        dimnames(var1) <- sub.dimensions[dims[as.character(order)]]
         dim.length[v] <- length(order)
         nc[[v]] <- var1
     }