+++ NEWS.md
@@ -27,6 +27,7 @@ functionality in `Derivate()`
 or fill scale. 
 - Fix for unconnected contours in `stat_contour_fill()`
 - New `geom_relief()` for generating relief shading. 
+- `ReadNetCDF()` gains a `subset` argument for subseting data.  
 
 # meteoR 0.0.9100
 - Full (not perfect) documentation.